Split service patcher
