+++ pkg/openstack/loadbalancer_service_patcher.go
@@ -1,0 +1,52 @@
+/*
+Copyright 2023 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package openstack
+
+import (
+	"context"
+	"reflect"
+
+	corev1 "k8s.io/api/core/v1"
+	utilerrors "k8s.io/apimachinery/pkg/util/errors"
+	"k8s.io/client-go/kubernetes"
+	cpoutil "k8s.io/cloud-provider-openstack/pkg/util"
+)
+
+type servicePatcher struct {
+	kclient kubernetes.Interface
+	base    *corev1.Service
+	updated *corev1.Service
+}
+
+func newServicePatcher(kclient kubernetes.Interface, base *corev1.Service) servicePatcher {
+	return servicePatcher{
+		kclient: kclient,
+		base:    base.DeepCopy(),
+		updated: base,
+	}
+}
+
+// Patch will submit a patch request for the Service unless the updated service
+// reference contains the same set of annotations as the base copied during
+// servicePatcher initialization.
+func (sp *servicePatcher) Patch(ctx context.Context, err error) error {
+	if reflect.DeepEqual(sp.base.Annotations, sp.updated.Annotations) {
+		return err
+	}
+	perr := cpoutil.PatchService(ctx, sp.kclient, sp.base, sp.updated)
+	return utilerrors.NewAggregate([]error{err, perr})
+}
